chore: move `docker_run` to `crate::near::docker_build`
