+++ cargo-near-build/src/lib.rs
@@ -18,12 +18,6 @@ pub use types::near::build::input::BuildContext;
 pub use types::near::build::input::{CliDescription, Opts as BuildOpts};
 // used in `AbiOpts` and `BuildOpts`
 pub use types::near::color_preference::ColorPreference;
-// TODO: remove on docker logic moved
-#[cfg(feature = "docker")]
-pub use near::docker_build::{docker_checks, git_checks, subprocess_step, ERR_REPRODUCIBLE};
-// TODO: remove on docker logic moved
-#[cfg(feature = "docker")]
-pub use types::near::docker_build as docker_build_types;
 #[cfg(feature = "docker")]
 pub use types::near::docker_build::Opts as DockerBuildOpts;
 
@@ -42,3 +36,5 @@ pub use camino;
 pub use near::abi::build as build_abi;
 pub use near::build::run as build;
 pub use near::build_extended::run as build_extended;
+#[cfg(feature = "docker")]
+pub use near::docker_build::run as docker_build;
