+++ cargo-near-build/src/near/docker_build/git_checks/dirty.rs
@@ -3,7 +3,8 @@ use std::time::Duration;
 use colored::Colorize;
 use eyre::{ContextCompat, WrapErr};
 
-use crate::{camino, docker_build_types::WARN_BECOMES_ERR, BuildContext};
+use crate::types::near::docker_build::WARN_BECOMES_ERR;
+use crate::{camino, BuildContext};
 use serde_json::to_string;
 
 pub fn check_then_handle(
